refactor: split recovery package into application domain and infrastructure
