+++ src/main/java/com/codenames/codenames/backend/recovery/application/SystemStateRecoveryService.java
@@ -1,4 +1,4 @@
-package com.codenames.codenames.backend.recovery;
+package com.codenames.codenames.backend.recovery.application;
 
 import com.codenames.codenames.backend.game.application.GameService;
 import com.codenames.codenames.backend.game.domain.GameManager;
@@ -6,7 +6,8 @@ import com.codenames.codenames.backend.game.domain.GameManagerFactory;
 import com.codenames.codenames.backend.lobby.domain.Lobby;
 import com.codenames.codenames.backend.lobby.api.dto.PlayerDto;
 import com.codenames.codenames.backend.lobby.application.LobbyService;
-import com.codenames.codenames.backend.recovery.snapshot.SystemSnapshot;
+import com.codenames.codenames.backend.recovery.domain.snapshot.SystemSnapshot;
+import com.codenames.codenames.backend.recovery.infrastructure.JsonStateStore;
 import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import java.util.Comparator;
 import java.util.List;
