+++ src/test/java/com/codenames/codenames/backend/recovery/application/SystemStatePersistenceServiceTest.java
@@ -1,4 +1,4 @@
-package com.codenames.codenames.backend.recovery;
+package com.codenames.codenames.backend.recovery.application;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
@@ -9,7 +9,8 @@ import static org.mockito.Mockito.when;
 import com.codenames.codenames.backend.game.application.GameService;
 import com.codenames.codenames.backend.lobby.api.dto.PlayerDto;
 import com.codenames.codenames.backend.lobby.application.LobbyService;
-import com.codenames.codenames.backend.recovery.snapshot.SystemSnapshot;
+import com.codenames.codenames.backend.recovery.domain.snapshot.SystemSnapshot;
+import com.codenames.codenames.backend.recovery.infrastructure.JsonStateStore;
 import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import java.util.List;
 import java.util.Map;
