+++ src/test/java/com/codenames/codenames/backend/recovery/application/SystemStateRecoveryServiceTest.java
@@ -1,4 +1,4 @@
-package com.codenames.codenames.backend.recovery;
+package com.codenames.codenames.backend.recovery.application;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -16,7 +16,8 @@ import com.codenames.codenames.backend.lobby.domain.Lobby;
 import com.codenames.codenames.backend.lobby.api.dto.PlayerDto;
 import com.codenames.codenames.backend.lobby.application.LobbyCodeGenerator;
 import com.codenames.codenames.backend.lobby.application.LobbyService;
-import com.codenames.codenames.backend.recovery.snapshot.SystemSnapshot;
+import com.codenames.codenames.backend.recovery.domain.snapshot.SystemSnapshot;
+import com.codenames.codenames.backend.recovery.infrastructure.JsonStateStore;
 import com.codenames.codenames.backend.game.dto.CardDataTransferObject;
 import com.codenames.codenames.backend.game.mapping.DataTransferObjectService;
 import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
